% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataplyr.R
\docType{package}
\name{odataplyr-package}
\alias{odataplyr}
\alias{odataplyr-package}
\title{This is a package}
\description{
Generate tbls for EntitySets
Generate functions for functions
Generate tbl_vars for Entity
}

